Add test which runs trinity for 90 seconds
