+++ tests/integration/test_trinity_cli.py
@@ -10,7 +10,10 @@ from eth.constants import (
     GENESIS_PARENT_HASH,
 )
 
-from trinity.tools.async_process_runner import AsyncProcessRunner
+from tests.integration.helpers import (
+    run_command_and_detect_errors,
+)
+
 from trinity._utils.async_iter import (
     contains_all
 )
@@ -24,25 +27,12 @@ from trinity._utils.async_iter import (
 # However, UNIX IPC paths can only be 100 chars which means long paths
 # *WILL* break these tests. See: https://unix.stackexchange.com/q/367008
 
-# This fixture provides a tear down to run after each test that uses it.
-# This ensures the AsyncProcessRunner will never leave a process behind
-@pytest.fixture(scope="function")
-def async_process_runner(event_loop):
-    runner = AsyncProcessRunner(
-        # This allows running pytest with -s and observing the output
-        debug_fn=lambda line: print(line)
-    )
-    yield runner
-    try:
-        runner.kill()
-    except ProcessLookupError:
-        pass
 
 # Great for debugging the AsyncProcessRunner
 # @pytest.mark.asyncio
 # async def test_ping(async_process_runner):
 #     await async_process_runner.run(['ping', 'www.google.de'])
-#     assert await contains_all(async_process_runner.iterate_stdout(), ['byytes from'])
+#     assert await contains_all(async_process_runner.stdout, {'bytes from'})
 
 
 @pytest.mark.parametrize(
@@ -168,20 +158,4 @@ async def test_web3(command, async_process_runner):
 async def test_does_not_throw(async_process_runner, command):
     # This is our last line of defence. This test basically observes the first
     # 20 seconds of the Trinity boot process and fails if Trinity logs any exceptions
-    lines_since_error = 0
-    await async_process_runner.run(command, timeout_sec=20)
-    async for line in async_process_runner.stderr:
-
-        # We detect errors by some string at the beginning of the Traceback and keep
-        # counting lines from there to be able to read and report more valuable info
-        if "Traceback (most recent call last)" in line and lines_since_error == 0:
-            lines_since_error = 1
-        elif lines_since_error > 0:
-            lines_since_error += 1
-
-        # Keep on listening for output for a maxmimum of 100 lines after the error
-        if lines_since_error >= 100:
-            break
-
-    if lines_since_error > 0:
-        raise Exception("Exception during Trinity boot detected")
+    await run_command_and_detect_errors(async_process_runner, command, 20)
